+++ src/controllers/estudiante.controller.ts
@@ -2,28 +2,45 @@ import { Request, Response } from "express";
 import { EstudianteService } from "../services/estudiante.service";
 import { HttpResponse } from "../shared/response/http.response";
 import { AuthAdminService } from "../auth/services/auth.service";
+import { GuardService } from "../auth/guards/verified.guard";
 
 export class EstudianteController{
     constructor(
         private readonly estudianteService: EstudianteService = new EstudianteService(), 
         private readonly httpResponse: HttpResponse = new HttpResponse(),
-        private readonly auth:AuthAdminService = new AuthAdminService()
+        private readonly auth:GuardService = new GuardService()
     ){   }
 
     async getEstudiantes(req: Request, res: Response) {
         try {
             
-            const data = await this.estudianteService.findAll();
-            if (data.length === 0) {
-                return this.httpResponse.NotFound(res, "No existen datos")
+            // Verificamos si esta logeado o no 
+            const token = req.cookies.accessToken
+            if (!token) {
+                return this.httpResponse.Unauthorized(res,"Inicie sesion primero");
             }
-            return this.httpResponse.Ok(res, data)
+
+            const typeUser = await this.auth.setPermissions(token);
+
+            if (typeUser === 1 || typeUser === 2) {
+                const data = await this.estudianteService.findAll();
+                if (data.length === 0) {
+                    return this.httpResponse.NotFound(res, "No existen datos")
+                }
+                return this.httpResponse.Ok(res, data)
+                
+            }
+
+            
+            return this.httpResponse.Unauthorized(res,"No tiene permiso para esta accion");
+
+
         } catch (e) {
             return this.httpResponse.NotFound(res, e)
         }
     }
 
-    async getEstudianteById(req: Request, res: Response) {
+    /* async getEstudianteById(req: Request, res: Response) {
         const {id}= req.params;
         try {
             
@@ -35,19 +52,30 @@ export class EstudianteController{
         } catch (e) {
             return this.httpResponse.NotFound(res, e)
         }
-    }
+    } */
 
     async createEstudiante(req: Request, res: Response) {
         try {
-            const data = await this.estudianteService.create(req.body);
-
-            if (data) {
-                return this.httpResponse.Ok(res, data)
-            }else{
-                return this.httpResponse.UserExists(res,"El usuario ya existe")
+            // Verificamos si esta logeado o no 
+            const token = req.cookies.accessToken
+            if (!token) {
+                return this.httpResponse.Unauthorized(res,"Inicie sesion primero");
             }
 
-            return this.httpResponse.Ok(res, data)
+            const typeUser = await this.auth.setPermissions(token);
+
+            if (typeUser === 1 || typeUser === 2) {
+                
+                const data = await this.estudianteService.create(req.body);
+    
+                if (data) {
+                    return this.httpResponse.Ok(res, data)
+                }else{
+                    return this.httpResponse.UserExists(res,"El usuario ya existe")
+                }
+            }
+            return this.httpResponse.Unauthorized(res,"No tiene permiso para esta accion");
+
         } catch (e) {
             return this.httpResponse.NotFound(res, e)
         }
@@ -56,13 +84,25 @@ export class EstudianteController{
     async updateEstudiante(req: Request, res: Response) {
         const {id}= req.params;
         try {
-            const data = await this.estudianteService.update(id,req.body);
-            
-            if (!data.affected) {
-                return this.httpResponse.NotFound(res, "Hay un error al actualizar")
+            // Verificamos si esta logeado o no 
+            const token = req.cookies.accessToken
+            if (!token) {
+                return this.httpResponse.Unauthorized(res,"Inicie sesion primero");
             }
 
-            return this.httpResponse.Ok(res, data)
+            const typeUser = await this.auth.setPermissions(token);
+
+            if (typeUser === 1 || typeUser === 2) {
+            
+                const data = await this.estudianteService.update(id,req.body);
+                
+                if (!data.affected) {
+                    return this.httpResponse.NotFound(res, "Hay un error al actualizar")
+                }
+    
+                return this.httpResponse.Ok(res, data)
+            }
+            return this.httpResponse.Unauthorized(res,"No tiene permiso para esta accion");
         } catch (e) {
             return this.httpResponse.NotFound(res, e)
         }
@@ -71,13 +111,26 @@ export class EstudianteController{
     async deteleEstudiante(req: Request, res: Response) {
         const {id}= req.params;
         try {
-            const data = await this.estudianteService.delete(id);
-            
-            if (!data.affected) {
-                return this.httpResponse.NotFound(res, "Hay un error al actualizar")
+            // Verificamos si esta logeado o no 
+            const token = req.cookies.accessToken
+            if (!token) {
+                return this.httpResponse.Unauthorized(res,"Inicie sesion primero");
             }
 
-            return this.httpResponse.Ok(res, data)
+            const typeUser = await this.auth.setPermissions(token);
+
+            if (typeUser === 1) {
+
+                const data = await this.estudianteService.delete(id);
+                
+                if (!data.affected) {
+                    return this.httpResponse.NotFound(res, "Hay un error al eliminar")
+                }
+    
+                return this.httpResponse.Ok(res, data)
+            }
+            
+            return this.httpResponse.Unauthorized(res,"No tiene permiso para esta accion");
         } catch (e) {
             return this.httpResponse.NotFound(res, e)
         }
@@ -94,6 +147,12 @@ export class EstudianteController{
     async findEstudianteWithSede(req: Request, res: Response) {
         const {id}= req.params;
         try {
+            // Verificamos si esta logeado o no 
+            const token = req.cookies.accessToken
+            if (!token) {
+                return this.httpResponse.Unauthorized(res,"Inicie sesion primero");
+            }
+
             const data = await this.estudianteService.findEstudianteWithSede(id);
             if (!data) {
                 return this.httpResponse.NotFound(res, "No existe datos")
